DHCP relaying issue fixed
The DHCP driver was storing the option 54 'DHCP Server Identifier' as
the senders IP address instead of from the option fields within a DHCP
offer.  This meant that DHCP relaying would not function correctly as
the gateway and DHCP server addresses are different in this instance.
The option field is now stored and used in the response.